backend/requirements.txt
asgiref==3.8.1
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Django==5.0.7
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-jalali==7.1.0
django-js-asset==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fonttools==4.53.1
fuzzywuzzy==0.18.0
geographiclib==2.0
jalali_core==1.0.0
jdatetime==5.0.0
kiwisolver==1.4.5
Levenshtein==0.26.0
matplotlib==3.9.2
numpy==2.0.1
packaging==24.1
pillow==10.4.0
PyJWT==2.8.0
pyparsing==3.1.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-Levenshtein==0.26.0
pytz==2025.2
qrcode==7.4.2
RapidFuzz==3.10.0
shapely==2.0.5
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
